postgresql: migration to Zabbix 3.4
This change breaks compatibility with Zabbix versions before 3.4 due
to the use of dependent items. With dependent items bringing up to a
factor of 10 load reduction on the agent I believe this is worth, given
that most users looking up these templates are likely coming from a
fresh install so having 3.4 or even 4.0 LTS.

Due to the export function of Zabbix reordering the XML file it's
nearly impossible to diff the template against the old version.
Sorry for that.

It brings as initial changes:
- filter database Low-Level-Discovery by macro. So what to ignore can
  be configured on a host level
- start to use dependent items to reduce load on the agent
- Don't rely on 8192 being the PostgreSQL block_size. Use the actual
  configured value. See #40
- I did not verify PostgreSQL compatibility with old versions, but bump
  the minimal to the oldest officially supported version 9.3