Update iostat-collect.sh
hello! I saw errors in file iostat-collect.sh: iostat man tels, that right order of options is "interval" and "count", not "count" and "interval". For delta statistic in N seconds it must be set count = 2, not count = 1 (in that wat it will be mean statistic since reboot). I hope this changes will be useful :) Thanks for great article on habr!