iostat: monitor stats collector
The template depends on a cronjob fetching and providing iostat
data in the background. As this is independent of the agent, it
requires monitoring, as it could break.

This change adds a timestamp to the statistics output. Time is
formatted as ISO to prevent issues with different locales. An
additional item in Zabbix stores the last received timestamp and
triggers a warning in case it differs more than 5 minutes from
server time.

A trigger means that either the cronjob has a problem or time
is out of sync.